// src/utils/enums/table_names.enum.ts
export enum TableNames {
  BroadCast = 'broadCast',
  Chat = 'chat',
  Patient = 'patient',
  Antenatal_visit = 'record_anc_antenatal_visit',
  Optional_Investigation = 'record_anc_optional_investigation',
  Antenatal_Care = 'record_antenatal_care',
  Pregnancy_overview = 'record_pregnancy_overview',
  Care_of_baby = 'record_care_of_baby',
  Baby_care = 'record_cob_baby_care',
  Postpartum_care = 'record_pnc_postpartum_care',
  Postnatal_care = 'record_post_natal_care',
  Prescription_rx = 'record_prescription_rx',
  Prescription = 'record_prescription',
  Screening_Form = 'record_screening_form',
  Lab_investigations = 'record_sf_lab_investigations',
  Vaccination_form = 'record_vaccination_form',
  Vital_form = 'record_vital_form',
  Template_rx = 'template_rx',
  Template = 'template',
  User = 'user',
  Visit = 'visit',
}